# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Locadora de Carros",
        "description": "Documentação da API da Locadora de Carros com autenticação JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/v1/marca": {
            "get": {
                "tags": [
                    "Marca"
                ],
                "summary": "Marca cadastradas",
                "operationId": "856be07665a618c27155ff131dd4c1df",
                "parameters": [
                    {
                        "name": "atributos_modelos",
                        "in": "query",
                        "description": "Quais colunas do relacionamento 'modelo' retornar. Use vírgula para separar.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "nome,imagem,marca_id"
                    },
                    {
                        "name": "filtro",
                        "in": "query",
                        "description": "Filtros dinâmicos. Exemplos de operadores comuns: =, >, >=, <, <=, like, in, between. ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "nome:like:Fiat"
                    },
                    {
                        "name": "atributos",
                        "in": "query",
                        "description": "Quais colunas do recurso 'locacao' retornar. Use vírgula para separar.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "nome,imagem,id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os dados da marca"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Marca"
                ],
                "summary": "Cadastrar marca",
                "operationId": "10534f7993cbe2f5fc7813bdc6e7af63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "imagem"
                                ],
                                "properties": {
                                    "nome": {
                                        "description": "Nome da marca (único).",
                                        "type": "string",
                                        "minLength": 3,
                                        "example": "Tesla"
                                    },
                                    "imagem": {
                                        "description": "Arquivo de imagem (formatos permitidos: png, jpg, svg, webp).",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Marca criada com sucesso"
                    },
                    "422": {
                        "description": "Erros de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/marca/{id}": {
            "get": {
                "tags": [
                    "Marca"
                ],
                "summary": "Detalhes de uma marca",
                "operationId": "513a38e6939b1dd5d608d8453267163e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da marca a ser pesquisado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Marca não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Marca"
                ],
                "summary": "Remover uma marca",
                "operationId": "836713a3cde497bfc9b88a4b6eb855fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da marca a ser deletada ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Marca não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT no formato: Bearer {seu_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}